+++ source/isaaclab_rl/isaaclab_rl/rsl_rl/__init__.py
@@ -21,3 +21,4 @@ from .rl_cfg import *
 from .rnd_cfg import RslRlRndCfg
 from .symmetry_cfg import RslRlSymmetryCfg
 from .vecenv_wrapper import RslRlVecEnvWrapper
+from .utils import handle_deprecated_rsl_rl_cfg
